requirements.txt
grpcio-tools~=1.50.0
tqdm~=4.64.1
numpy~=1.21
